requirements.txt
baidu-aip==2.2.18.0
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
Pillow==7.1.1
PyMuPDF==1.16.18
PySimpleGUI==4.18.2
requests==2.23.0
urllib3==1.25.9
